extract the test consistency manager
